Fix exported frames share sheet bug
Frames were being prematurely deleted in some cases during a share activity.

This happened when an activity completes without dismiss the share sheet, typically when an activity was cancelled. It also seems to happen with Shortcuts in general.

+++ Frame Grabber/Editor/EditorViewController.swift
@@ -293,10 +293,19 @@ private extension EditorViewController {
 
     func share(urls: [URL]) {
         let shareController = UIActivityViewController(activityItems: urls, applicationActivities: nil)
-        shareController.completionWithItemsHandler = { [weak self] _, _, _, _ in
+
+        shareController.completionWithItemsHandler = { [weak self] activity, completed, _, _ in
+            guard self?.shouldDeleteFrames(after: activity, completed: completed) == true  else { return }
             self?.videoController.deleteExportedFrames()
         }
+
         presentOnTop(shareController)
+    }
+
+    func shouldDeleteFrames(after shareActivity: UIActivity.ActivityType?, completed: Bool) -> Bool {
+        let wasDismissed = (shareActivity == nil) && !completed
+        let didFinish = (shareActivity != nil) && completed
+        return wasDismissed || didFinish
     }
 
     // MARK: Showing Progress
